pyproject.toml
[project]
name = "routix"
version = "0.0.12"
description = "Define routines as structured algorithmic workflows"
readme = "README.md"
authors = [
    { name = "JuneTech", email = "junetech0@gmail.com" }
]
license = "MIT"
requires-python = ">=3.11"
dependencies = []

[project.urls]
Repository = "https://github.com/MSOlab/routix.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "types-pyyaml>=6.0.12.20250402",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
